# testdata/openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorResponse": {
        "description": "Error response",
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Monitor": {
        "description": "A heartbeat monitor (dead man's switch)",
        "example": {
          "enabled": true,
          "grace_period_seconds": 1800,
          "id": "019c0123-4567-7890-abcd-ef1234567890",
          "inserted_at": "2026-01-29T10:00:00Z",
          "interval_seconds": 86400,
          "last_ping_at": "2026-02-06T02:05:00Z",
          "name": "Nightly Backup",
          "next_expected_at": "2026-02-07T02:05:00Z",
          "ping_token": "pm_abc123def456",
          "ping_url": "https://runlater.eu/ping/pm_abc123def456",
          "schedule_type": "interval",
          "status": "up",
          "updated_at": "2026-02-06T02:05:00Z"
        },
        "properties": {
          "cron_expression": {
            "description": "Cron expression (for cron monitors)",
            "nullable": true,
            "type": "string"
          },
          "enabled": {
            "default": true,
            "description": "Whether the monitor is active",
            "type": "boolean"
          },
          "grace_period_seconds": {
            "default": 300,
            "description": "Grace period before alerting (seconds)",
            "type": "integer"
          },
          "id": {
            "description": "Monitor ID",
            "format": "uuid",
            "type": "string"
          },
          "inserted_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "interval_seconds": {
            "description": "Expected interval in seconds (for interval monitors)",
            "nullable": true,
            "type": "integer"
          },
          "last_ping_at": {
            "description": "When the last ping was received",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "description": "Monitor name",
            "type": "string"
          },
          "next_expected_at": {
            "description": "When the next ping is expected",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "notify_on_failure": {
            "description": "Override org-level failure notification setting. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "notify_on_recovery": {
            "description": "Override org-level recovery notification setting. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "ping_token": {
            "description": "Unique ping token",
            "type": "string"
          },
          "ping_url": {
            "description": "Full ping URL",
            "format": "uri",
            "type": "string"
          },
          "schedule_type": {
            "description": "Schedule type",
            "enum": [
              "cron",
              "interval"
            ],
            "type": "string"
          },
          "status": {
            "description": "Current monitor status",
            "enum": [
              "new",
              "up",
              "down",
              "paused"
            ],
            "type": "string"
          },
          "updated_at": {
            "description": "Last update timestamp",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "schedule_type"
        ],
        "title": "Monitor",
        "type": "object"
      },
...
      "MonitorRequest": {
        "description": "Request body for creating or updating a monitor",
        "example": {
          "grace_period_seconds": 1800,
          "interval_seconds": 86400,
          "name": "Nightly Backup",
          "schedule_type": "interval"
        },
        "properties": {
          "cron_expression": {
            "description": "Required for cron monitors",
            "nullable": true,
            "type": "string"
          },
          "enabled": {
            "default": true,
            "type": "boolean"
          },
          "grace_period_seconds": {
            "default": 300,
            "description": "Grace period in seconds (0-3600)",
            "type": "integer"
          },
          "interval_seconds": {
            "description": "Required for interval monitors (60-604800)",
            "nullable": true,
            "type": "integer"
          },
          "name": {
            "description": "Monitor name",
            "type": "string"
          },
          "notify_on_failure": {
            "description": "Override org-level failure notification setting. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "notify_on_recovery": {
            "description": "Override org-level recovery notification setting. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "schedule_type": {
            "enum": [
              "cron",
              "interval"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "schedule_type"
        ],
        "title": "MonitorRequest",
        "type": "object"
      },
      "MonitorResponse": {
        "description": "Response containing a monitor",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Monitor"
          }
        },
        "title": "MonitorResponse",
        "type": "object"
      },
      "MonitorsResponse": {
        "description": "Response containing a list of monitors",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Monitor"
            },
            "type": "array"
          }
        },
        "title": "MonitorsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "API key in format: pk_live_xxx.sk_live_yyy",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for managing scheduled tasks and webhooks.\n\n## Authentication\n\nAll API requests require a Bearer token in the Authorization header:\n\n```\nAuthorization: Bearer pk_live_xxx.sk_live_yyy\n```\n\nAPI keys can be created in the Runlater dashboard under Organization Settings.\n",
    "title": "Runlater API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/monitors": {
      "get": {
        "callbacks": {},
        "description": "Returns all monitors for the authenticated organization",
        "operationId": "PrikkeWeb.Api.MonitorController.index",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorsResponse"
                }
              }
            },
            "description": "Monitors list"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all monitors",
        "tags": [
          "Monitors"
        ]
      },
      "post": {
        "callbacks": {},
        "description": "Creates a new heartbeat monitor",
        "operationId": "PrikkeWeb.Api.MonitorController.create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorRequest"
              }
            }
          },
          "description": "Monitor parameters",
          "required": false
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorResponse"
                }
              }
            },
            "description": "Monitor created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a monitor",
        "tags": [
          "Monitors"
        ]
      }
    },
    "/api/v1/monitors/{id}": {
      "delete": {
        "callbacks": {},
        "description": "Deletes a monitor and all its ping history",
        "operationId": "PrikkeWeb.Api.MonitorController.delete",
        "parameters": [
          {
            "description": "Monitor ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Monitor deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a monitor",
        "tags": [
          "Monitors"
        ]
      },
      "get": {
        "callbacks": {},
        "description": "Returns a single monitor by ID",
        "operationId": "PrikkeWeb.Api.MonitorController.show",
        "parameters": [
          {
            "description": "Monitor ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorResponse"
                }
              }
            },
            "description": "Monitor"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a monitor",
        "tags": [
          "Monitors"
        ]
      },
      "patch": {
        "callbacks": {},
        "description": "Updates an existing monitor",
        "operationId": "PrikkeWeb.Api.MonitorController.update (2)",
        "parameters": [
          {
            "description": "Monitor ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorRequest"
              }
            }
          },
          "description": "Monitor parameters",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorResponse"
                }
              }
            },
            "description": "Monitor updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a monitor",
        "tags": [
          "Monitors"
        ]
      },
      "put": {
        "callbacks": {},
        "description": "Updates an existing monitor",
        "operationId": "PrikkeWeb.Api.MonitorController.update",
        "parameters": [
          {
            "description": "Monitor ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorRequest"
              }
            }
          },
          "description": "Monitor parameters",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorResponse"
                }
              }
            },
            "description": "Monitor updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a monitor",
        "tags": [
          "Monitors"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://runlater.eu",
      "variables": {}
    }
  ],
...
}